Add NeighborConnectivity check
If the HTTP header NeighborConnectivity is set, the function receiving
the request will attempt to reach all other HTTP functions and report
which are reachable. The corresponding HTTP request to probe for the
connectivity has the NoOperation header set which causes no additional
work performed on the receiving side of the request.